[Feature] 상점 BigInt 구현

+++ Assets/ScriptableObjects/ClickSO.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "ClickSO", menuName = "ScriptableObjects/ClickSO", order = 1)]
 public class ClickSO : ScriptableObject
 {
-    public int ClickValue = 1;
+    public long ClickValue = 1;
     public float AutoClickInterval = 1f;
     public bool AutoClickEnabled = false;
 }
